requirements.txt
anyascii==0.3.2
asgiref==3.7.2
beautifulsoup4==4.11.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2.11
django-axes==6.4.0
django-crum==0.7.9
django-filter==23.5
django-modelcluster==6.3
django-otp==1.3.0
django-permissionedforms==0.1
django-taggit==4.0.0
django-treebeard==4.7.1
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
gunicorn==21.2.0
html5lib==1.1
idna==3.6
l18n==2021.3
mysqlclient==2.1.1
openpyxl==3.1.2
packaging==24.0
pillow==10.2.0
pillow_heif==0.15.0
pyasn1==0.5.1
pycparser==2.21
pypng==0.20220715.0
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
typing_extensions==4.10.0
urllib3==2.2.1
wagtail==5.2.3
wagtail-2fa==1.6.9
wagtail-modeladmin==2.0.0
webencodings==0.5.1
Willow==1.6.3
